+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.slice 0.1.1.9014.2
+# teal.slice 0.1.1.9015
 
 * Removed `CDISCFilteredDataset` class and functionality moved to `CDISCFilteredData`.
 * Changed constructor of `FilteredData` to not require `TealData` object. See `help(init_filtered_data)` for more details.